+++ src/lib/schema/index.ts
@@ -1,0 +1,4 @@
+import formLoginSchema from "@/lib/schema/loginSchema";
+import formSignUpSchema from "@/lib/schema/signUpSchema";
+
+export { formLoginSchema, formSignUpSchema }
